coastwatch-ai/backend/requirements.txt
# PelicanEye - Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pillow>=10.2.0
ultralytics>=8.1.0
pydantic>=2.5.0
aiofiles>=23.2.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
